requirements.txt
confluent-kafka==2.3.0
requests==2.31.0
boto3==1.34.0
snowflake-connector-python==3.7.0
python-dotenv==1.0.0
prefect==2.16.0
dbt-snowflake==1.7.0
streamlit==1.32.0
pandas==2.2.0
